requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-compat==1.0.15
django-dotenv==1.4.2
djongo==1.3.6
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.0.1
pyparsing==3.0.6
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.7
